only speculate on binops with typefeedback
so far we assumed non objects on types which have no typefeedback. This
leads to deopt-loops in the tests, so I changed it.

but to avoid regressions, we need to become better at using
typefeedback. As a first step I propagate the typefeedback when we
replace instructions. Ie. if we replace a for b, I copy the typefeedback
of a to b. This might not always be correct, but seems like a decent
hack for the kind of transformations we are doing mostly.