More constant fold (#440)
Added constant folding for Eq and Neq when one of the arguments is a simple logical and the other is a simple scalar, constant folding for Not when the arg is a simple scalar or double negation (although double negation technically wouldn't be a constant fold, should it go in cleanup?), and constant folding for AsInt when the arg is a simple real or logical.

These particularly help simple ranges. The latter allows the old version which normally has 2 loops to remove the second loop in cases like for (i in 1:10), and the other optimizations remove some dead branches / simplify the current version which always uses 1 loop.