compile while(...) {} as if(...) do {} while (...)
this is specifically to give us a preheader to force promises used in
the loop condition. E.g. in:

    function(x)
        ...
        while (i < x) { ... }

our optimizations fail, because `x` is forced in a loop. But if we write
the equivalent

    function(x)
        ...
        if (i < x)
          do { ... } while (i < x)

It's very easy to see where x is forced and also a lot simpler to inline
that promise and to determine that the second x is not forcing at all.