extend scope analysis to support merging of undeclared values
If we have the case:

     e = env()
    loop:
     checkpoint ...
     stvar "x", 42, e
     goto loop

and want to move "e" into the deopt case, we need to create an env,
which is `env(x=phi(unboundValue,42))`, since on the first iteration
of the loop the value is not declared, but on the second iteration it
is.

This PR adds support for exactly that case.