Loop detection code (#478)
* Loop detection code

Does the standard textbook approach of identifying loops. A loop is
defined by a header h, and a back edge n->h where h dominates n. If
there are multiple back edges to the same header, the body is the union
of all the nodes computed for each back edge.

Once back edges are identified, we do a reverse depth-first search from
the "tail" until we reach the header.

There is an option to compute the loop-nesting forest, but it disabled
by default because it is slow.