Removing dead phis needs to account for cycles
Before this commit, a phi that refers to itself is considered live.
Furthermore, a phi referred to by a dead phi would also be considered
live. Therefore, we have to construct a dependency graph and then do a
DFS traversal -- if we reach a non-phi instruction then the phi is
alive. Otherwise, nothing live refers to the phi, so it can be removed.