don't remove checkpoint if no next checkpoint is available
Given this case:

    checkpoint 1
      |
      ... (a)
      |
    checkpoint 2
      |
      ... (b)
      |
      effect
      |
    checkpoint 3

Then cp1 is available at cp2 (ie. all instructions at (b) can also use cp1
instead of cp2) and it seems we can remove it. But for elide_env_spec we
need the next checkpoint. If we remove cp2 then (a) does not have a next
cp anymore, since cp3 is blocked by an effect. Thus we should only
remove cp2 if there is also a next cp available.