Use call_ instead of call_implicit_ in rir
This commit is the first step to deprecate call_implicit_. The reason is
that creating explicit promises and using call_ allows us to deoptimize
between promise creation and call_. In case we discover that the call_
goes to a builtin, we can eagerly evaluate arguments. To be able to have
checkpoints during eager argument evaluation however we need checkpoints
between creating promises and the actual call. This is not possible with
the call_implicit_ instruction, which creates promises and does the call
in one instruction. Thus the plan to deprecate it. Currently call_
cannot deal with dots arguments, this needs to be addressed before we
can completely remove it.