Fix bug with constantfold
This is actually a subtle bug that arises from the complicated CFGs
generated by loop peeling.

When constant folding sees that a branch condition is constant, it can
delete one of the branches. For example, say the true branch is deleted.
The algorithm will then go and delete all BBs dominated by the true
branch.

However, consider a more complicated CFG:

          A     B
         / \   / \
            \ /
             C

If the branches to C are dead code, then C should be deleted. However, A
and B do not individually dominate C, so our algorithm does not delete
C.

Instead, we should look at A and B as a set, and compute the BBs
dominated by that set. This commit implements that algorithm.

Pair programmed with @o-.