Change the visitor used by phi placement
The CFG is visited and a currently valid value for a given variable is propagated through it. However, using the dominator tree visitor could sometimes lose track in case there was a diamond shape with multiple BBs along the paths, such as
    BB1
   /  \
 BB2  BB3
  |    |
 BB4  BB5
   \  /
    BB6
Here the visitor would enter BB1 and then queue BB2, BB3 and BB6, because all those are immediately dominated by BB1. Then BB2 and BB3 would see the tracked value from BB1, but BB6 wouldn't as it only becomes visible for it after visiting BB4 or BB5.
Using BFS should fix this.