add preliminary pir support for `...`
This commit adds 3 new instructions:

* LdDots
* DotsList
* ExpandDots

LdDots is like a `ldvar ...`, except that it does not error when it
is missing. It returns a DOTSXP or missing.

DotsList creates a DOTSXP list from name/value pairs.

ExpandDots is the most magic one. It expands a DOTSXP to a number of
arguments.

At the type level we differ between `dots` which is a DOTSXP list and
`*dots` (created by ExpandDots) which is an expanded DOTSXP list.
When a `*dots` argument is passed to a call, this corresponds to an
arbitrary number of arguments (including 0). For example a call
`f(...)` translates to `Call(f, ExpandDots(LdDots()))`.

Arguments of type `*dot` cannot be passed to a `StaticCall`, since
there we need to know the exact number and names of arguments.
Arguments of type `dot` cannot be passed to a normal `Call`, since
the normal call does dynamic argument name matching, which expects
the arguments to be individually expanded (creating the dotslist
happens implicitly while argument matching).

Here is how those instructions get used:

1) If we call a function which has `...` as a formal argument, then
we can convert this call to a `StaticCall` by statically matching
the argument names and using `DotsList` to combine arguments
captured by `...`. For example in:

    f = function(a, ..., b) {g(a,b,...)}
    f(c=1,b=2,a=3, 4)

The call translates to

    d = dotsList(c=1, 4)
    StaticCall(f, {3, d, 2})

2) If we pass `...` as an argument, we can use the expandDots to
convert a dotsxp to a number of args. In the above example the
call to `g` translate to:

    e = ExpandDots(LdDots())
    Call(g, {LdVar(a), LdVar(b), e})

3) Inlining works with arguments of type `...`. After inlining we see:

    d = dotsList(c=1, 4)
    env = MkEnv(a = 3, ...=d, b=2)
    e = ExpandDots(LdDots())
    Call(g, {LdVar(a), LdVar(b), e})

Therefore scope resolution kicks in:

    d = DotsList(c=1, 4)
    e = ExpandDots(d)
    Call(g, {3, 2, e})

Now `ExpandDots` and `DotsList` annihilate each other and this results
in:

    Call(g, {3, 2, c=1, 4})

At the rir level we add a new `mk_dotlist` bytecode, which corresponds to
the `DotsList` pir instruction. `ExpandDots` and `LdDots` is implicitly
handled by the existing `call_dots` bytecode.