fix memory corruption of the stubenv after it materialized
if a stub env is materialized then the context points to the
materialized env. but the `env` variable of the interpreter might still
point to the stub. therefore we need to ensure that the gc does not
collect the stub. this patch just keeps materialized stubs alive
forever. we need a more reasonable fix, potentially by just removing the
env variable from the interpreter.