don't treat error as a deopt barrier
the reasoning is that errors terminate the function, so no rollback
before the error can happen, thus they can be ignored for assumes.

also this fixes a bug where the assume opt pass would remove too
many cp's since it did skip the first instr of the next bb when
serching for a next cp.

debugging and thinking by @charig