fix eager call pass
the pass was missing an upcase, which caused:

   Force(MkArg(...))

Since MkArg is eager (it returns a promise-as-a-value), the force
is a noop, and the whole thing returns a promise. This caused
promises to flow into eager builtins.

Instead we need:

   Force(CastType(lazy, MkArg(...)))

With this the promise-as-a-value is castet to a lazy-value and
force actually evaluates it.