speculate on all ldvars
Speculate on ldvar and ldvar_super results.

in the past we didn't do that out of fear that we would introduce bogus
checks. but instead this PR adds some constant folding to remove bogus
checks should they appear later. for example this can happen if the
scope resolution pass resolved the ldvar later.

also fix a bug where availableCheckpoint.next returns a checkpoint that
is not dominated by the requested position. that is not good, since we
cannot add tests to such a checkpoint.