rewrite the static named count approximation
the previous analysis was mostly broken. In particular it did not
correctly detect cases where values overwritten by subassign are use.
For example:

    subassign(..., x, ...)
    use(x)

when translated naively to rir, will cause x to be overwritten by
subassign and then used later in modified form, when actually ssa
semantics would demand that it is unmodified.

the new analysis is based on a taint analysis, that tries to detect
values that are tainted and then used after that.

it improves the precision by correctly modeling how values flow into
phis and how phis get tainted or not.