speculate on vectors where we might unboxed extract
this will speculate on types such as `(int|real)"<int">`, because
such a speculation will let us use an unboxing `Extract`.

In particular this targets expressions such as
`for (i in 1:b)` where often `1:b` is an int vector but we don't
statically know, since it could also be a float vector
(actually, can it?).