ensure there are checkpoints when forcing arguments
the cast insertion pass cannot insert checkpoints. since it may insert
side-effecting instructions, we end up with side effecting instructions
without checkpoint in between. long term we should get rid of it completely.
for now this PR ensures that at least the force instructions are inserted
eagerly and with checkpoints in between.