remove useless analysis of promises in deadstore
the recursive analysis of promises does not contribute to the precision,
since we still fall into the generic case. instead they slow down this
analysis massively. instead the promise analysis was just masking a bug
in the leak analysis, since we should query the leak analysis before the
call and not after the call (otherwise the envs leaked through the call
are not considered for the effects of the call).