Improved recording of types
Previously our type feedback mechanism would record
(attribute,object,scalar) flags for each type individually. When the
number of types overflows this leaves no other option than to assume
these are tainted. To improve precision we move the flags out of the
individual types, such that even if the SEXPTYPES overflow, we still
keep track precisely of the attributes.

pair-programmed with @skrynski