ensure that when moving environments the context position is updated
if we move an environment over a PushContext we need to update its
context offset field, or it will register itself to the wrong context.
E.g.

    e1 = mkenv context 1
         PushContext ... e2
         Checkpoint ....

if we move `e1` into the checkpoint, then the context must be increased
to 2, otherwise it will override the environment `e2` in the inlined
context.