Fix arglist reordering metadata in the presence of ...
The promargs list recovery after statically rewriting it was completely
broken for `...` args. It did not consider that `...` can occur
multiple times in the arguments. It did also not consider that some
`DOTSXP` need to be flattened to get the original args list (because we
statically rewrote them) but others have to stay `DOTSXP`, because the
user passed them that way. E.g. I learned that you can access the
dots using `environment()['...']`.

This commit ensures that we record the fact if `...` was expanded in the
argument reordering metadata and only expand it based on that.