refactor deopt reason
This PR is a preparation to more intelligent deopts. For that we remove
the RecordDeoptReason instruction. Instead the deopt reason becomes a
PIR value that can be passed around and passed to the deopt instruction.

Also this patch removes ScheduledDeopt in favor of just handling the
normal Deopt instruction in the backend. To that end some better
handling of value liveness was needed.

While refactoring I also got rid of automatic conversion between
Representation and llvm::Type. This makes some code less confusing.