Fix naked dots miscompilation
In rir, we use R_DotsSymbol for marking an argument that should be expanded at runtime, but also when we need to push the actual symbol (in quote(...)).
In pir we don't know anymore which case this was, and so try to always expand the dots.
This adds a symbol placeholder .expandDotsTrigger. to be used in the arglists instead.