fix compilation of nested loops
When we have nested simple loops, then a break or continue in the body
of an inner loop caused the loops to have contexts. The reason is that
when we compile the fallback case for the simple loop, then we forgot to
set up a loop context. This means that the break being compiled for
fallback would taint the wrong (ie. outer) loop context.