Don't dispatch to native code that's still being compiled (#1203)
* Don't dispatch to native code that's still being compiled

In the native backend a Code object could be accessed from a DispatchTable while still being compiled
(we call deparse to print pir constants which can eval code). Such a Code would wrongly think it's bytecode
and start interpreting garbage.

This adds a pending flag that prevent dispatching to function versions that have pending Code objects. The flag
is set when we create a native Code object and reset when we finalize the LLVM module, at which point native
code can be emitted.

* address review, split bc/native in rir::Code