Support deopts in outermost promises
This adds support for deopting from promises that are the outermost frame when deopting.
The mechanism is that if the outermost frame is a promise, we don't longjump to a context
(since promises don't have contexts), but rather return the result. This is the only case
when deopt returns, and we immediately return also from the promise code with the result,
passing it to the place where the promise forcing occured.