Remove extra contexts while deopting (#1222)
* cleaning up

* more cleanup

* Remove extra contexts while deopting

Because of an optimization where we replace the first checkpoint in an inlinee by the checkpoint at the callsite it is possible to take the outer deopt exit from within the inlinee. In case the inlinee needs its own context we need to pop it, otherwise the context offsets in MkEnv instructions in the deopt block would modify the wrong frame.
```
a = checkpoint -> [deopt with (mkenv ... context 1)]
...
c = push_context ....
assume ..., a -> we have to pop c
```
The other option would be to make the context field in MkEnvs a PIR variable and then set it to the correct value based on where we come from, however, this seems easier.