Fix verifier choking when inlining a promise in the RIR bytecode compiler
`function() invisible(return(5))` would trip the bytecode verifier since we'd emit ret_ in the inlined promise in the middle of the `invisible` call sequence, leaving stuff on the stack...
For now, track when inlining promises and emit return_ in those cases (since return_ does a longjump and resets stack).
Ideally we would just not emit any code after finding a call to return, but this would be more involved to change.