remove old R serialization, R serialization now uses the new serializer API.
It's redundant enough, and we've already been comparing it with the new RIR serialization and there have been no discrepancies except for function stats which are still being investigated. So instead of doing what was done with hashing and getConnected where we have the old version (which may even be kept permanently due to performance), just remove and replace with a wrapper to call RIR's serializers.

Also removed some redundancies and an old TODO which isn't necessary (tryReadHash and tryWriteHash seems better off in UUIDPool)